Move FancyAlertComponent to dedicated file in UITestsFoundation
