--- WordPress/WordPressUITests/Screens/FancyAlertComponent.swift
@@ -1,1 +1,0 @@
-// FIXME: Delete once transition is completed
